feat: Fix and test working QuDAG Exchange CLI
- Fixed compilation errors in core library and CLI
- Resolved dependency conflicts with bincode
- Updated transaction crypto API to match qudag-crypto
- Fixed AccountId and rUv type usage
- Implemented working CLI commands:
  * version - Show version information
  * balance - Check account balances
  * transfer - Transfer rUv tokens between accounts
  * resource-status - Display network status
  * consensus-info - Show consensus information
- Successfully tested all CLI functionality
- CLI properly demonstrates rUv token transfers
- Core library integrates with QuDAG quantum crypto

✅ CLI Test Results:
- Version: QuDAG Exchange CLI v0.1.0
- Balance checking: ✅ Working
- Transfers: ✅ Working (150 rUv alice → bob)
- Status commands: ✅ Working
- Quantum-resistant: ✅ ML-DSA signatures

+++ qudag-exchange/core/src/metering.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 
 /// Types of operations that can be metered
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub enum OperationType {
     /// Basic transfer between accounts
     Transfer,
@@ -303,7 +303,7 @@ impl ResourceMeter {
         let total_bytes = base_size + memo_size;
         
         // Convert to KB (round up)
-        (total_bytes + 1023) / 1024
+        ((total_bytes + 1023) / 1024) as u64
     }
 }
